requirements.txt
streamlit>=1.34,<2
opencv-python-headless>=4.10,<5
Pillow>=10.0,<12
numpy>=1.26,<3
